+++ src/components/shared/layout/LayoutHeaderDesktop.tsx
@@ -4,14 +4,13 @@ import dynamic from 'next/dynamic'
 import Image from 'next/image'
 import Link from 'next/link'
 import { useRouter } from 'next/router'
-import { PiCodesandboxLogo, PiCoinsLight, PiChartLine, PiGlobe  } from 'react-icons/pi'
+import { PiCodesandboxLogo, PiCoinsLight, PiChartLine, PiGlobe } from 'react-icons/pi'
 import styled from 'styled-components'
 import stLogoDesktop from '../../../../public/assets/stake-together-desk.svg'
 import useActiveRoute from '../../../hooks/useActiveRoute'
 import useLocaleTranslation from '../../../hooks/useLocaleTranslation'
 import SkeletonLoading from '../icons/SkeletonLoading'
-import Search from '@/components/search/SearchControl'
-
+import LayoutHeaderSearch from './LayoutHeaderSearch'
 const LayoutNetworkDropdown = dynamic(() => import('./LayoutNetworkDropdown'), {
   ssr: false,
   loading: () => <SkeletonLoading width={80} height={32} />,
@@ -57,7 +56,7 @@ export default function LayoutHeader() {
             </MenuButton>
           </Link>
         </Menu>
-        <Search />
+        <LayoutHeaderSearch />
       </MenuContainer>
       <WalletContainer>
         <LayoutNetworkDropdown />
@@ -72,7 +71,19 @@ export default function LayoutHeader() {
   )
 }
 
-const { Container, MenuContainer, WalletContainer, ContainerCurrency, Logo, Menu, MenuButton, GlobeIcon, ProjectsIcon, CoinsIcon, ChartIcon } = {
+const {
+  Container,
+  MenuContainer,
+  WalletContainer,
+  ContainerCurrency,
+  Logo,
+  Menu,
+  MenuButton,
+  GlobeIcon,
+  ProjectsIcon,
+  CoinsIcon,
+  ChartIcon
+} = {
   Container: styled.header`
     display: none;
     gap: ${({ theme }) => theme.size[32]};
@@ -127,7 +138,6 @@ const { Container, MenuContainer, WalletContainer, ContainerCurrency, Logo, Menu
       white-space: nowrap;
       cursor: pointer;
     }
-  
   `,
   Menu: styled.nav`
     display: none;
